lotto-62 Moved the lotto_app/app/commands/command_utils.py file to lotto_app/app/management/command_utils.py. Created Draft command_8add command.
