+++ lotto_app/app/management/command_utils/command_utils.py
