+++ lotto_app/app/management/commands/command_8add.py
